Many changes:
Upon creating a main function and attempting to add elements to the circuit board, I found that the "id" attribute was not being initialized. This was because the derived class contructor was not initializing the "id" attribute in the base class constructor.
This was fixed.

However, the issue now is that I have to make the attributes in the active components (magnitude, type, multiplier) accessible, as they are not with my design. This is because each element is a BoardComponent type and not an active component or node.