A million changes.
First, I changed the actual data objects (Board Components) to have variables more accessible later on in the code.
Plans are made to use encapsulation and getter methods, but as of now are made public for personal convenience.

Changed the names to match more Electrical Engineering terminology.

Also, added all attributes to the Base class. This, IMO, nullifies the point of inheritance at this point, as each object (data wise) is practically the same. The only big difference is that Wire automatically initializes each magnitude and multiplier to 0.
Power supply and passive compnonents are the same, IMO.

Changes are made to make them have functions that have the same name but different operations.
The main function is mind is when throwing errors and what would cause them.

For example, one should not call a power supply a resistor, but this would be a fine thing to call a passive component.
Things like this will be addressed in the virtual functions.

On the functional side, made it so one can see the system of equations that the user created circuit has.

Note: this system has a slew of errors, but the skeleton has been made.

Errors include:
1) does not account for adjacent voltage and current sources when calculating node voltages.
2) Is only a string representation, does not actually numerically compute anything.
3) Does not account for reference node (and thus has 1 extra equation).
4)Does not account for Super Nodes